Move project formatting tests to a separate file
