+++ src/XMakeBuildEngine/UnitTests/Evaluation/Preprocessor_Tests.cs
@@ -832,191 +832,5 @@ namespace Microsoft.Build.UnitTests.Preprocessor
 
             Helpers.VerifyAssertLineByLine(expected, writer.ToString());
         }
-
-        //  TODO: Test preprocessor formatting (project.SaveLogicalProject)
-        //  TODO: If preserveFormatting is true, don't change single quotes to double quotes
-        //  TODO: Add tests for preserving formatting when mutating project (ie adding an item, modifying a property)
-
-        //  TODO: Not sure the formatting tests actually belongs in the preprocessor tests
-
-        [Fact]
-        public void ProjectCommentFormatting()
-        {
-            string content = ObjectModelHelpers.CleanupFileContents(@"
-<Project DefaultTargets=`Build` ToolsVersion=`msbuilddefaulttoolsversion` xmlns=`msbuildnamespace`>
-  <ItemGroup>
-    <ProjectReference Include=`..\CLREXE\CLREXE.vcxproj`><!-- Comment -->
-      <Project>{3699f81b-2d03-46c5-abd7-e88a4c946f28}</Project>
-    </ProjectReference>
-  </ItemGroup>
-</Project>");
-
-            string reformattedContent = ObjectModelHelpers.CleanupFileContents(@"
-<Project DefaultTargets=`Build` ToolsVersion=`msbuilddefaulttoolsversion` xmlns=`msbuildnamespace`>
-  <ItemGroup>
-    <ProjectReference Include=`..\CLREXE\CLREXE.vcxproj`>
-      <!-- Comment -->
-      <Project>{3699f81b-2d03-46c5-abd7-e88a4c946f28}</Project>
-    </ProjectReference>
-  </ItemGroup>
-</Project>");
-
-            VerifyFormattingPreserved(content);
-            VerifyProjectReformatting(content, reformattedContent);
-        }
-
-        [Fact]
-        public void ProjectWhitespaceFormatting()
-        {
-            string content = ObjectModelHelpers.CleanupFileContents(@"
-<Project DefaultTargets=`Build` ToolsVersion=`msbuilddefaulttoolsversion` xmlns=`msbuildnamespace`>
-  <ItemGroup>  
-    <ProjectReference Include=`..\CLREXE\CLREXE.vcxproj`>
-<Project>{3699f81b-2d03-46c5-abd7-e88a4c946f28}</Project>
-    </ProjectReference>
-  </ItemGroup>
-</Project>");
-
-            string reformattedContent = ObjectModelHelpers.CleanupFileContents(@"
-<Project DefaultTargets=`Build` ToolsVersion=`msbuilddefaulttoolsversion` xmlns=`msbuildnamespace`>
-  <ItemGroup>
-    <ProjectReference Include=`..\CLREXE\CLREXE.vcxproj`>
-      <Project>{3699f81b-2d03-46c5-abd7-e88a4c946f28}</Project>
-    </ProjectReference>
-  </ItemGroup>
-</Project>");
-
-            VerifyFormattingPreserved(content);
-            VerifyProjectReformatting(content, reformattedContent);
-        }
-
-        [Fact]
-        public void ProjectQuoteFormatting()
-        {
-            string content = ObjectModelHelpers.CleanupFileContents(@"
-<Project DefaultTargets='Build' ToolsVersion='msbuilddefaulttoolsversion' xmlns='msbuildnamespace'>
-  <ItemGroup>
-    <ProjectReference Include='..\CLREXE\CLREXE.vcxproj'>
-      <Project>{3699f81b-2d03-46c5-abd7-e88a4c946f28}</Project>
-    </ProjectReference>
-  </ItemGroup>
-</Project>");
-
-            string reformattedContent = ObjectModelHelpers.CleanupFileContents(@"
-<Project DefaultTargets=""Build"" ToolsVersion=""msbuilddefaulttoolsversion"" xmlns=""msbuildnamespace"">
-  <ItemGroup>
-    <ProjectReference Include=""..\CLREXE\CLREXE.vcxproj"">
-      <Project>{3699f81b-2d03-46c5-abd7-e88a4c946f28}</Project>
-    </ProjectReference>
-  </ItemGroup>
-</Project>");
-
-            VerifyFormattingPreserved(content);
-            VerifyProjectReformatting(content, reformattedContent);
-        }
-
-        void VerifyFormattingPreserved(string projectContents)
-        {
-            VerifyFormattingPreservedFromString(projectContents);
-            VerifyFormattingPreservedFromFile(projectContents);
-        }
-
-        void VerifyFormattingPreservedFromString(string projectContents)
-        {
-            ProjectRootElement xml = ProjectRootElement.Create(XmlReader.Create(new StringReader(projectContents)),
-                ProjectCollection.GlobalProjectCollection,
-                preserveFormatting: true);
-            Project project = new Project(xml);
-            StringWriter writer = new StringWriter();
-            project.Save(writer);
-
-            string expected = @"<?xml version=""1.0"" encoding=""utf-16""?>" +
-                projectContents;
-            string actual = writer.ToString();
-
-            Helpers.VerifyAssertLineByLine(expected, actual);
-        }
-
-        void VerifyFormattingPreservedFromFile(string projectContents)
-        {
-            string directory = null;
-
-            try
-            {
-                directory = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
-                Directory.CreateDirectory(directory);
-
-                string file = Path.Combine(directory, "test.proj");
-                File.WriteAllText(file, projectContents);
-
-                ProjectRootElement xml = ProjectRootElement.Open(file, ProjectCollection.GlobalProjectCollection,
-                    preserveFormatting: true);
-                Project project = new Project(xml);
-                StringWriter writer = new StringWriter();
-                project.Save(writer);
-
-                string expected = @"<?xml version=""1.0"" encoding=""utf-16""?>" +
-                    projectContents;
-                string actual = writer.ToString();
-
-                Helpers.VerifyAssertLineByLine(expected, actual);
-            }
-            finally
-            {
-                Directory.Delete(directory, true);
-            }
-        }
-
-        void VerifyProjectReformatting(string originalContents, string expectedContents)
-        {
-            VerifyProjectReformattingFromString(originalContents, expectedContents);
-            VerifyProjectReformattingFromFile(originalContents, expectedContents);
-        }
-
-        void VerifyProjectReformattingFromString(string originalContents, string expectedContents)
-        {
-            ProjectRootElement xml = ProjectRootElement.Create(XmlReader.Create(new StringReader(originalContents)),
-                ProjectCollection.GlobalProjectCollection,
-                preserveFormatting: false);
-            Project project = new Project(xml);
-            StringWriter writer = new StringWriter();
-            project.Save(writer);
-
-            string expected = @"<?xml version=""1.0"" encoding=""utf-16""?>" +
-                expectedContents;
-            string actual = writer.ToString();
-
-            Helpers.VerifyAssertLineByLine(expected, actual);
-        }
-
-        void VerifyProjectReformattingFromFile(string originalContents, string expectedContents)
-        {
-            string directory = null;
-
-            try
-            {
-                directory = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
-                Directory.CreateDirectory(directory);
-
-                string file = Path.Combine(directory, "test.proj");
-                File.WriteAllText(file, originalContents);
-
-                ProjectRootElement xml = ProjectRootElement.Open(file, ProjectCollection.GlobalProjectCollection,
-                    preserveFormatting: false);
-                Project project = new Project(xml);
-                StringWriter writer = new StringWriter();
-                project.Save(writer);
-
-                string expected = @"<?xml version=""1.0"" encoding=""utf-16""?>" +
-                    expectedContents;
-                string actual = writer.ToString();
-
-                Helpers.VerifyAssertLineByLine(expected, actual);
-            }
-            finally
-            {
-                Directory.Delete(directory, true);
-            }
-        }
     }
 }
