+++ src/XMakeBuildEngine/UnitTestsPublicOM/Microsoft.Build.Engine.OM.UnitTests.csproj
@@ -78,6 +78,7 @@
     <Compile Include="Construction\ConstructionEditing_Tests.cs" />
     <Compile Include="Construction\ElementLocationPublic_Tests.cs" />
     <Compile Include="Construction\ProjectChooseElement_Tests.cs" />
+    <Compile Include="Construction\ProjectFormatting_Tests.cs" />
     <Compile Include="Construction\ProjectImportElement_Tests.cs" />
     <Compile Include="Construction\ProjectImportGroupElement_Tests.cs" />
     <Compile Include="Construction\ProjectItemDefinitionGroupElement_Tests.cs" />
